application: Added extra config hook.
This allows customization of the blueprint before it's added to the application.
Recent versions of Flask started printing out a warning, saying that I should
not be modifying, then re-registering blueprints which was what the consumer of
this library was doing and now it doesn't need to do that anymore if it uses the
post-processing hook.