Switched to feedwerk for atom feeds.
This removes the dependency on lxml which has a weird build system. The main
goal was to make it possible to distribute an application embedding gitpages as
a single file distribution. lxml was one of the most difficult to distribute.
The next is actually docutils which is even more difficult due to lots of
__file__ usage within the code. Not sure that will ever get fixed.